Move parser directory under controller
Updated relavent paths for manager classes.
