+++ src/main/java/seedu/duke/CcaManager.java
@@ -3,7 +3,7 @@ package seedu.duke;
 import seedu.duke.common.Messages;
 import seedu.duke.exception.CcaEmptyStringException;
 import seedu.duke.exception.CcaParamException;
-import seedu.duke.parser.DateTimeParser;
+import seedu.duke.controller.parser.DateTimeParser;
 import seedu.duke.ui.UserInterface;
 
 import java.text.ParseException;
