+++ src/main/java/seedu/duke/controller/parser/DateTimeParser.java
@@ -1,4 +1,4 @@
-package seedu.duke.parser;
+package seedu.duke.controller.parser;
 
 import seedu.duke.LogManager;
